🧪 363FBC code: split sqlite cache regression test
